Moved the AuthController to testHelpers to remove it from the final build.
