+++ sources/packages/api/src/auth/roles.enum.ts
@@ -3,5 +3,4 @@
  */
 export enum Role {
   Student = "StudentRole",
-  DummyRole = "DummyRole",
 }
